tester/requirements.txt
speedtest-cli~=2.1.3
psycopg2-binary~=2.7